refactor: split main file into cogs;   - rename angel6-Rewrite.py to angel6.py   - Split it into [fun, logging, moderation, system, utility]   - Run Black --preview on all files
